% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_log_like.r
\name{approx_log_like}
\alias{approx_log_like}
\title{Evaluate the approximate log marginal likelihood function}
\usage{
approx_log_like(
  RespLog,
  par_val,
  long.data,
  surv.data,
  Bi,
  invSIGMA,
  distribution,
  degree
)
}
\description{
Evaluate the approximate log marginal likelihood function
}
