% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_base_hazard.r
\name{est_base_hazard}
\alias{est_base_hazard}
\title{Estimate the baseline hazard function in the Cox model as a step function}
\usage{
est_base_hazard(surv.data, survObject, par_val, Bi)
}
\description{
Estimate the baseline hazard function in the Cox model as a step function
}
