% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_disp.r
\name{est_disp}
\alias{est_disp}
\title{Estimates the dispersion parameters in the joint models, 
by maximizing the adjusted profile h-likelihood function.}
\usage{
est_disp(
  param,
  invSIGMA,
  other_param,
  RespLog,
  deriv_vars,
  long.data,
  surv.data,
  Bi,
  Lval0 = NULL,
  lower,
  upper,
  distribution = "normal",
  degree = 0,
  Silent = T
)
}
\description{
Estimates the dispersion parameters in the joint models, 
by maximizing the adjusted profile h-likelihood function.
}
