% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_fixeff_rob.r
\name{est_fixeff_rob}
\alias{est_fixeff_rob}
\title{Estimate fixed parameters using robust inference method}
\usage{
est_fixeff_rob(
  param,
  sigma,
  other_param,
  model_object,
  long.data,
  Bis,
  invSIGMA,
  Silent = TRUE
)
}
\description{
Estimate fixed parameters using robust inference method
}
