% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_randeff.r
\name{est_randeff}
\alias{est_randeff}
\title{Obtain estimates of random effects}
\usage{
est_randeff(RespLog, long.data, surv.data, subject_id, random_effects,
  invSIGMA, par_val, distribution = "normal", degree = 0, Silent = T,
  scale = T)
}
\description{
Obtain estimates of random effects
}
