% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_randeff.r
\name{est_randeff_by_hlike}
\alias{est_randeff_by_hlike}
\title{Estimate random effects by maximizing h-likelihood function,
given fixed and dispersion parameters}
\usage{
est_randeff_by_hlike(RespLog, sub_long, sub_surv, random_effects, invSIGMA,
  par_val, distribution, degree, Silent = T)
}
\description{
Estimate random effects by maximizing h-likelihood function,
given fixed and dispersion parameters
}
