% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_sigma_rob.r
\name{est_sigma_rob}
\alias{est_sigma_rob}
\title{Robust estimate of standard error of residual term}
\usage{
est_sigma_rob(param, lower = 0, upper = 100, fixed_param,
  other_param = NULL, model_object, long.data, Bis, Silent)
}
\description{
Robust estimate of standard error of residual term
}
