% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_disp.r
\name{eval_Hmat_for_disp}
\alias{eval_Hmat_for_disp}
\title{Evaluate H matrix}
\usage{
eval_Hmat_for_disp(Hmats, par_val, long.data, surv.data, B, Bi, invSIGMA,
  distribution, degree)
}
\description{
Evaluate H matrix
}
