% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_log_like.r
\name{eval_log_hlike}
\alias{eval_log_hlike}
\title{Evaluate log h-likelihood}
\usage{
eval_log_hlike(RespLog, par_val, long.data, surv.data, B, Bi, invSIGMA,
  distribution, degree)
}
\description{
Evaluate log h-likelihood
}
