% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHHJM.R
\name{fitHHJM}
\alias{fitHHJM}
\title{Fit robust joint model based on h-likelihood}
\usage{
fitHHJM(glmeObjects, survObject, long.data, surv.data, subject_id,
  randeff_info = list(distribution = "t-dist", degree = 3),
  itertol = 0.001, Ptol = 0.01, iterMax = 10, Silent = T,
  mcmc_size = 100, burn_in = 100, thin = 20)
}
\arguments{
\item{survObject}{a list that provides the necessary information of 
a survival model.}

\item{long.data}{longitudinal data}

\item{surv.data}{survival data}

\item{glmeObject}{a list of sublists. Each sublist provides the necessary
information of an LME or GLME model.}
}
\description{
This function is for robust joint modelling of multiple (generalized) 
linear mixed-effect models and a survival model (Cox PH or Weibull).
The continuous longitudinal data may be left-censored due to lower 
limit of quantification.
}
