% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_density_glm.R
\name{get_log_density_glm}
\alias{get_log_density_glm}
\title{Get log density function of (generalized) linear model}
\usage{
get_log_density_glm(model_object)
}
\description{
Get log density function of (generalized) linear model
}
