% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_density_ph.R
\name{get_log_density_ph}
\alias{get_log_density_ph}
\title{Get log density function of propotional hazard model}
\usage{
get_log_density_ph(model_object)
}
\description{
Get log density function of propotional hazard model
}
