% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sd_MCEM.r
\name{get_sd_MCEM}
\alias{get_sd_MCEM}
\title{Estimate s.e. of non-robust fixed parameters,
using MCEM approximation}
\usage{
get_sd_MCEM(param, RespLog, long.data, surv.data, Bis, random_effects,
  subject_id, other_param, distribution, degree, Silent)
}
\description{
Estimate s.e. of non-robust fixed parameters,
using MCEM approximation
}
