% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{getHmat}
\alias{getHmat}
\title{Derive H matrix in profile h-likelihood functions}
\usage{
getHmat(RespLog, pars)
}
\description{
Density of random effects is excluded here and will 
be included separately while estimating parameters.
Note that this function returns -H.
}
