% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_randeff_MCMC.r
\name{likelihood}
\alias{likelihood}
\title{Log h-likelihood of longitudinal and survival data}
\usage{
likelihood(xx, par_list, sub_long, sub_surv)
}
\description{
Log h-likelihood of longitudinal and survival data
}
