% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_randeff_MCMC.r
\name{posterior}
\alias{posterior}
\title{Un-normalized log posterior of random effects}
\usage{
posterior(param, par_list, sub_long, sub_surv)
}
\description{
Un-normalized log posterior of random effects
}
