% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{strMat}
\alias{strMat}
\title{This function returns a matrix with elements in string.
Moreover, it returns L(l)'L(l) = SIGMA, which is a spherical
parameterization with diagonal elements equal to 1.}
\usage{
strMat(q2)
}
\description{
This function returns a matrix with elements in string.
Moreover, it returns L(l)'L(l) = SIGMA, which is a spherical
parameterization with diagonal elements equal to 1.
}
