% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{Vderiv}
\alias{Vderiv}
\title{Compute derivatives of an expression w.r.t 
a vector of parameters respectively}
\usage{
Vderiv(lik1, pars)
}
\description{
Compute derivatives of an expression w.r.t 
a vector of parameters respectively
}
